重构完成，还需要重构测试用例
