+++ src/content-scripts/server.js
@@ -1,8 +1,9 @@
 /**
- * 内容脚本同时也是一个 Server 端，用来报告自己的 location 给扩展程序
+ * 内容脚本同时也是一个 Server 端，用来执行扩展程序发送过来的命令
  */
 
 import {Server} from 'connect.io';
+import getST from './init-st';
 
 const server = new Server();
 
@@ -11,6 +12,13 @@ server.on( 'connect' , client => {
     if ( self === top ) {
       resolve( JSON.parse( JSON.stringify( location ) ) );
     }
+  } );
+
+  client.on( 'translate' , ()=> {
+    getST().then( st => {
+      st.query.text = getSelection().toString();
+      st.translate();
+    } );
   } );
 } );
